Split process discovery out of the quiescence module
